return empty structure
Before we would return nil, meaning no table would be displayed.
Since we always want to display the table, even if it's empty, we need to
return an empty structure.

Note:
- I had to add the team.default_language so the first row would be displayed,
and not only the header.

Tests:
- I added a new test to make sure we are returning all the columns
- I updated the older tests to check if Month has an hyphen, instead of
data_report being nil. This still makes sure we are starting from an
empty data_report.