Make sure that `null` is not returned as the item when creating an imported fact-check with original media in GraphQL API. (#2153)
When handling an existing media to decide if the imported fact-check is going to be added or appended to the existing media, there is a case where the media already exists with a fact-check in the same language that is trying to be imported. That case was returning nil. This PR fixes it by making sure that the new object is returned in that case, and then let the validations take care of returning a persisted or failed object.

Fixes: CV2-5804.