5857 – apply_replace_by should save history version and not error if the old_pm does not exist (#2163)
When we run apply_replace_by, we look for the original project media id, but sometimes that doesn't exist anymore.

This only causes an error when we try to save that information to the history versioning. Since we already have access to that original project media id, we can just grab that value and save that information.

Note: We are not sure why that happens, but there are places in that area of the code that expect the possibility of old_pm being nil. So for now, we just update to get the information we need and avoid this specific error.

References: 5857
PR: 2163