Bug – CV2-5913 – Cannot update a new record (#2173)
We had an error (ActiveRecord::ActiveRecordError: cannot update a new record) only that happened in a very specific context:
- we have an existing ProjectMedia with a LinkMedia
- we make a graphql request with:
      - the same link in set_original_claim
      - the same language
      - the title field is requested as part of the response

This happened because of the way we create_link_media and handle_fact_check_for_existing_claim(existing_pm, new_pm).

The solution was change handle_fact_check_for_existing_claim to not just return new_pm (since that was returning an object that was not persisted) but to try to save! it. So we either return a persisted object or an error is raised

CV2-5913, PR 2173