add tests
Right now the tests still fail, because we haven't added the new behavior.
Some notes:

- Currently we are still creating the media from the original_claim through
ProjectMedia, so I'm adding the tests to ProjectMedia.
- I'll add the new behavior, make sure the tests are working, and then
work on refactoring: we want to move this from ProjectMedia to Media.
– On the last test, I think, there is some extra weirdness going on:
it is actually creating 4 Medias: 2 Claims and 2 Links. I don't think it
should create the Link Medias. I'll look into that.