Adding new arguments to `TeamType.bot_preview` GraphQL field. (#2183)
We already have the `TeamType.bot_query` GraphQL field implemented, and it takes only one argument (the text query). It currently uses the settings from the workspace and/or Smooch Bot installation. We want to make sure that it can receive additional optional arguments that override the stored settings.

Reference: CV2-5919.