rename so it's clear media is created from original_claim
We have very similar methods in ProjectMediaCreators.
I think it probably makes sense to move those to Media
as well, and try to deduplicate.

But for now, so there is no confusion, I want to make
it very clear that we are using these methods when
creating from original_claim.