update Tests
We don't need to declare the hash, a callback does that for us.
That said I'm unsure about which tests to keep in this file.
Since we want to make sure this behavior works when creating medias
from project medias with original_claim, it makes sense to have a few
tests here. But I'm unsure of which.