[WIP] First part: move methods to Media
I moved the methods to Media so it's easier to compare them side-by-side,
and to make sure this works from Media.

I also made a small change to the create_media method in ProjectAssociation,
so its behavior is closer to create_original_claim. Since those two have differences,
but are both basically creating a Media associated to the ProjectMedia.
Hopefully those two will be turned into one method.