trying to combine methods and remove set_media_type
One big difference between how we create Media is in one case we were
sending the entire project_media, but when creating from original_claim
we weren't.

I think since this is Media it makes more sense only to receive what we
need, instead of the whole project_media. So this is a step in that
direction.

There is still a lot of duplication, and I think there should be better
ways to access the values we want.