Fixing the media origin for an item added to a media cluster by the Smooch Bot (#2241)
The media origin for a media item added to a media cluster by the Smooch Bot was set to "user merged" when it should have been "auto merged," since the Smooch Bot is, after all, a bot. The fix expands the "auto merged" case to include not only the Alegre Bot but any bot.

Fixes: CV2-6205.