Delete the `Comment` annotation type (#2219)
* Delete the `Comment` annotation type

The “comment” (a.k.a. “notes”) feature was deprecated a while ago,
and it’s not exposed on the frontend side anymore. It means that we
can delete it from the backend as well.

* Remove references to `Comment` in /test

Remove all references to `Comment` and `create_comment` from our tests.

* Fix remaining tests

* Remove more references to `Comment`  from tests

* Revert change to create_anotation, replacing _comment with _tag

* Add reviewer feedback

* Update GraphQL relay schema

Update GraphQL relay schema

* Remove more references to Comment from tests

Remove more references to Comment from tests

* Fix test based on reviewer feedback

* Fix code coverage issues

Fix code coverage issues

* Cover uncovered lines in team.rb

* Cover uncovered lines

* Revert removed method

* CV2-6063: add missing tests

---------

Co-authored-by: Sawy <melsawy.be7eri@gmail.com>