check for nil and empty string (#2248)
In a request set_original_claim might not be present at all, or it might
be present, but be an empty string. So we need to check for both of those.

I added set_original_claim: "" to a test, whioch makes sure we are now
testing for this as well.