4916 – Change 'listed' attribute so we do not display Fetch (#2250)
Fetch is deprecated and should not be shown in the workspace. We have already removed Fetch specific code from Check-Web. But it still renders the generic parts, if the api sends it.

We have a boolean attribute called listed. Which was implemented here: https://github.com/meedan/check-api/commit/d0cd2747bc035aebbd33a9bae3bf60fc5ebe15b8

Listed bots are the ones that are displayed to users so they can install in their workspaces.

Since we don't want to show Fetch we can update it's listed attribute to "false", so it won't be displayed anymore. Note that Fetch listed attribute is already set to false in check live. We want it to be set t false regardless of environment.

References: CV2-4916