(5995) Implement timeout and fallback for `original_claim` URL processing (#2243)
Context
The current implementation for processing original_claim URLs in Check API does not account for situations where external URLs take an extended time to process. This can result in delays or failures if processing exceeds the gateway timeout (currently set at 60 seconds).

This mostly happens when we try to create an UploadedFile Media from an original_claim. So we see this a lot in our Zapier integrations, where we might get the app took too long to respond. This happens somewhat frequently, and could obscure real issues.

How
    - We configured a Timeout (shorter than 60s),
    - if that is exceeded, we create a ClaimMedia instead of Audio/Image/VideoMedia.

Notes
- find_or_create_uploaded_file_media
I realized we would sometimes send the file, and sometimes send the URL. Then the method would need to figure out what to do. With that implementation, we would need to add a begin/rescue to this method, so it could then call find_or_create_claim_media if the download failed.

    - I think we should always send the same thing to find_or_create_uploaded_file_media, so we should always send a file.
    - If ProjectMedia is downloading the file, it can decide from the start where to send it, instead of sending it to one method and that redirecting to another one.
    - So ProjectMedia tries to download the file, and decides which method to call,
    - find_or_create_uploaded_file_media and find_or_create_claim_media only worry about creating the Media

- find_or_create_uploaded_link_media
Unfortunately, right now we couldn't just deal with the timeouts in one place. For Links timeouts happen during the Link creation itself, while for files, it happens during download. To make sure Timeout is raised, we had to add a rescue and re-raise it in PenderData. (we can control the timeout window in the PenderClient gem)

References: CV2-5995