Tests for fact check
We are currently relying on User.current to set the channel, but
writing the test for this was really tricky, and I'm thinking might
point to a bigger issue.

It might also be an issue with how our methods from sample_data are
written. I'm not sure.

Keeping all of these for now so we can discuss.