rake task to migrate articles channel to api
We want to set channel to `api` if it was created by a BotUser,
if it was created by a human user we want to set it as manual.

Since we are defaulting channels to `manual` in this migration:
db/migrate/20250219110106_add_channel_to_explainers_and_fact_checks

We only need to update the cases where it was created by a Bot to `api`.