6319 – Update timeout when creating Media (#2279)
When we create a media instance through Zapier (original_claim) we often hit a Timeout, specially when creating Link media or File media instances. So for those cases we want our timeout to be shorter.

We added a global configuration key for the short timeout - for two reasons, first because we have it in two places, so it’s important to have a single source of truth, and second, because it will be easier for us to tweak that value if needed without needing to go through PR review, deployment, etc.

References: CV2-6319, CV2-5995