Only save the report sent timestamp when the report is actually sent.
Previously, the report sent timestamp was saved before the report was actually sent, which could result in a situation where a report was not sent but the information was incorrectly recorded as if it had been. This PR fixes the issue by ensuring that the report sent timestamp is stored in the tipline request only when, and after, the report is actually sent.

Fixes: CV2-6317.