add create mutation: first part
We create a webhook, which means a BotUser with specific attributes.
So a bot_user is returned. Notes:
– Not sure if we should have a wrapper to hide this and return a webhook?
– Because we return a bot_user, I had to implement the fields in BotUserType

I had to add def `request_url=(request_url)` and  `def headers=(headers)`
so graphql_crud_operations would successfully add the request_url and headers.

I still need to be able to return the webhook list once a webhook is created