destroy mutation: reproduce error
Ale saw an error from the frontend that happens because when
we query for the webhooks, it returns the encoded base 64 id as 'Webhook/ID',
instead od 'BotUser/ID', then when we try to use that value to delete
the bot_user/webhook we get `uninitialized constant Webhook\nDid you mean?  WebMock`.

This only updates the test to reproduce this behavior. It is not fixed yet.