try to restrict Dynamic,DynamicAnnotation::Field destroy by team
When we create the user, we create a Source associated to it.
However, when we were setting the team for source self.team_id and Team.current
were both nil. So we failed to set a team.

If we fail to set a team, we could try to get it from the associated user.
Then when trying to destroy it, we could try to check the team.
However #2, Dynamic does not seem to have a team associated to it?
We could grab the Source associated to that Dynamic ID, and check for
its team.

Notes:
- This looks convulated
– Not sure if there is side effects
– Is there a better way to do this?