5271 – 6369 – Refactor to make current feed list explicitly a media list (#2307)
Context:
First part of the work to add Articles as Available Data to Share when creating a shared feed. This is a refactoring to rename saved_search to media_saved_search. 

This was done because Feed and FeedTeam can store a single reference to a list (saved search), which is assumed to be a media list. We want Feeds to be able to reference media_saved_search and article_saved_search. We also had to make changes in the model and graphql layer.

This is a refactoring to keep the current behavior, so we didn't change how the API works. Nothing should change in the frontend. 

References: CV2-5271, CV2-6369
PR: 2307

Changes:
* Update Feed and FeedTeam column  saved_search to media_saved_search
* Update feed and feed_team belongs_to
* Update Feeds and FeedTeams model and unit tests
* Update GraphQL layer and controller tests
     - feed_type and feed_team_type: Specify method to get media_saved_search and media_saved_search_id, 
so we don't need to change how we request it from the frontend. I also added a test to make sure the request works. Doc: https://graphql-ruby.org/fields/introduction.html
    - feed_team_mutations and feed_mutations: Specify the argument name with as: :media_saved_search_id. Doc: https://graphql-ruby.org/fields/arguments.html