5271 – 6370 – Add articles saved search (#2310)
Before, shared feeds can only store one media saved search, this updates it so that we can store one media saved search, and/or one article saved search. As part of this work, we:

    - added an article_saved_search reference to Feed and FeedTeam
    - added validation of saved_search type
    - updated Feed DATA_POINTS to include 'articles' instead of 'published_fact_checks'
    - added new fields to FeedType and FeedTeamType
    - added new arguments to FeedMutation and FeedTeamMutation
    - update saved_search_was to specify whether it is media or article saved search
    - removes saved_search aliasing
    - updated relay
    - updated tests
    - updated seeds script

This work does not update how we display the lists in the shared feed. Only enables us to add, save, delete the lists when creating, updating the Feed.

This needs to be updated in the frontend as well, as this would break it: meedan/check-web#2364

References: CV2-5271, CV2-6370