deal with failling test
I ran into RuntimeError: No permission to update FeedTeam.
At first I just added authentication to the test, but after talking to
Caio I added feed_team.skip_check_ability = true to the update_feed_team,
and removed the authentication from the test.

This fixes the issue and is more consistent with the rest of the code.