6361 – Main feed's feed team should return main feed information (#2317)
Issue
When making a request to return the current user feed's feed teams, the saved searches are correct for the invited teams, but return nil for the owner feed (even if they exist in the main feed)

Solution
When we create a new Feed instance, an after_create callback runs and creates a FeedTeam instance with the main Feed instance. I think if we are going to create the main Feed's FeedTeam instance, then maybe we should add the saved searches when creating it. We should make sure the data between Feed and FeedTeam is consistent.

I also added a rake task to update existing data.

References: CV2-6361