Update test: We don't want to duplicate team bots
When we duplicate team bots, specially the ones used to access API keys,
we can run into issues. E.g.:
- We create a team with a bot that has an API key to import items
- We duplicate the team, which duplicates the bot and API key
- Now we might think we are importing items to the original team,
but we are actually importing them to the duplicated team

In order to avoid different teams with same API keys, instead of copying
the original team's bots, we want to just install any default bots.