skip_check_ability if team.is_being_copied
instead of adding a second argument to install_to to skip_check_ability,
we can only skip it if the team is being copied. This way we can
ensure that the bot installation is created without checking permissions
when duplicating a team, but still enforce permissions when installing
the bot to a team in other contexts.