6248 – teambotinstallation permissions issue (#2326)
Context
An integration test (should duplicate team) in Check-Web is not passing anymore.

{
    "message": "No permission to create TeamBotInstallation",
    "code": 5,
    "data": {}
}

Dani tested this manually and confirmed the behavior:
   - Super admin users can duplicate a team successfully.
   - Non-super admin users receive the same permission error as the test.

This indicates a permission issue related to the TeamBotInstallation when performed by non-super admin users.

TLDR: Non-super admin users should be able to duplicate teams, so we need to fix this permissions issue.

Solution
This was happening because we were in the context of the original team, and original team user, but the TeamBotInstallation belonged to the duplicate team.

Setting team_bot_installation.skip_check_ability = true in BotUser.install_to!, if the team is being copied.
This way we can ensure that the bot installation is created without checking permissions when duplicating a team, but still enforce permissions when installing the bot to a team in other contexts.

Relates to: #2323, CV2-3942
References: CV2-6428