Remove inotify-tools dependency and sidekiq script
We used inotify to monitor the restart.txt file and
restart Sidekiq in development mode, when using
touch tmp/restart.txt to restart rails.

The way we were doing this is not supported in
Sidekiq 6, and we want to simplify our Dockerfiles
foe development and production so we can get them
more similar.

If needed, we can restart Sidekiq by manually
restarting the container: docker compose restart api-background

In the future, we might want to explore using autoreloading
and features present in Sidekiq.