6507 – Make sure we look for an Account created from a profile link (#2345)
Issue
When we ran Bot::Keep.webhook we were failing to check for cases where Pender parsed a type: profile, meaning we always returned a 425 in those cases.

Example
   - I have Keep turned on,
   - I create a Media from this link https://x.com/user/status/000000000000
   - It also creates the Source Account https://x.com/user
   - It tries to archive both urls
   - But because we only look for Link, Pender keeps getting 425 for the Account
       - And while it does, it will keep retrying the Sidekiq job until all retries are exhausted

Solution
We check for Account as well, in which case we don't need to update Project Media annotations, and send back a simple response.

References: CV2-6507