Make sure tipline languages are present as supported languages
When using the setup_smooch_bot while passing menu = true, we add
two tipline workflows: 'en' and 'pt'.

But we were not installing 'pt' in the team.

For a language to be present in a tipline workflow, it needs
to be a supported language in the team.

This was causing tests to misbehave.