Stop running all tests twice (#2365)
We were running all tests twice because we are running on push and on pull_request.
If a test run fails for, let's say push, but succeeds for pull_request, we just ignored it and merge anyway. Besides, sometimes a run would even fail to start because of the multiple tests running.

We now run on pull requests against any branch, but on push only against master and develop.