Make sure we update the correct object
When we were updating the settings, s, to be a HashWithIndifferentAccess
in the apply_default_settings method, the settings object was not the
same as s anymore. We were not updating the object we expected.

We want to make sure we return the correct object, if there are any
other places in code where we are doing this as well.