Deal with ActionController::Parameters in CheckSettings (#2374)
We caused a bug with the work done in: 
https://github.com/meedan/check-api/pull/2369

In bot_user, we use the dynamic get methods with events.
So in those cases, the current_value we get in define_method is an
instance of ActionController::Parameters, with permitted parameters.

We apparently had no tests covering this, and while in QA this issue
did not surface, it did in live.

So here we deal with ActionController::Parameters, if the value is
permitted, we wrap it in a HashWithIndifferentAccess.

I also added tests for the CheckSettings library as a whole, and
added one that mimics the parameters we get in bot_user.

References: CV2-6581